requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
yookassa>=3.3.0